+++ package.json
@@ -47,6 +47,7 @@
     "zod": "^3.8.2"
   },
   "devDependencies": {
+    "@types/qrcode.react": "^1.0.2",
     "@typescript-eslint/eslint-plugin": "^4.31.0",
     "@typescript-eslint/parser": "^4.31.0",
     "eslint": "^7.32.0",
